[devshell] fx setup-ufw: new helper script
This CL adds a new devshell script to allow updating the Linux
firewall rules in order to allow Fuchsia-related traffic,
from devices and emulators, to go through.

This allows using them on systems with strict firewalls setups.
The rules only allow packets on IPv6 link-local interfaces and
should be safe in any environment (since it means the Fuchsia
device should be directly connected to the host computer).

+ Ensure that "fx run -N" and "fx aemu -N" will print a small
  warning to the user if an active ufw firewall is detected
  on Linux.

BUG: ONB-53
TEST: Tested manually on a corp laptop.
Change-Id: I54023ab2d1620433aea51ba2ee9c5d0e72beff85