[benchmarks] Run zircon_benchmarks via SL4F
This is an initial step towards using SL4F for running all of the
performance tests in fuchsia.git.  We start by running
zircon_benchmarks this way.  Add the Dart code for doing this, based
on existing tests such as src/tests/end_to_end/screen_is_not_black.

We split the zircon_benchmarks package into two packages to deal with
it being run in two modes.

Note that this does yet not run catapult_converter to convert the
results for uploading to the Catapult performance dashboard, because
that runs into a problem with using runtime_deps on the fuchsia.git
bots.  Running catapult_converter will be added in a later change.

For local testing:

  # Use host_asan to match the fuchsia.git bots:
  fx set terminal.x64 --args='select_variant=["host_asan"]'
  # fx build + re-pave
  fx run-e2e-tests perf_test

The new SL4F test is run on the "terminal-x64-release" trybot.
zircon_benchmarks is run in unit test mode on the main fuchsia bots.

Bug: 23091
Test: trybots mentioned above + local testing steps above
Change-Id: I983901a24a7a141a45d61ea26391e736d456e6bd