[fuchsia-hyper] Parse link-local IPv6 URL zones as the scope id
This adds support for hyper clients to communicating to link local
addresses. At the moment, we only interpret the URI zone to be an integer
scope, with plans to later (see bug 39402) to change fuchsia-hyper to talk
to the netstack in order to map from a string name to the appropriate scope
id.

Bug: 36623

Change-Id: I8b2aa81ca28260e2e7abc75c7431d73e7b3178ae