[camera][utils] Write images to disk
This CL is for a small util to read image data from a buffercollection
and write it to a persistent partition on disk.

WIP note: currently attempting to plug it into the camera demo flow.

There will be at least two more CL's that build on this:
* One that allows ImageIOUtil to "scp" the image from the device to the
host PC.
* Another that provides the SL4F Facade interface so the host PC can
interact with the device.

TESTED=`fx run-test camera_full_on_device_test -t stream_utils_test` and
demo test still passes

Change-Id: I5bd7e473f3d2801b86d9f78b5c432646b58da97d