[debugger] Format vector registers in expressions
Adds a target-specific setting to control vector register formatting.

Hook this register up to the expression evaluator to support printing a
vector:
  [zxdb] print xmm0
  {2.17, 3.14}
and extracting individual elements:
  [zxdb] print xmm0[1]
  3.14

This hook-up is done by a new EvalContext subclass in the client library
that integrates with the settings system.

The formatting is not yet hooked up to the "regs" command.

BUG=39308

Change-Id: I89ce6b53d1b2019e1cab74452b2d97aa87fc4481