[escher] Fix circular sempahore dependency in BatchGpuUploader
Fix a latent bug in the BatchGpuUPloader where if multiple readers
attempt to read from the same buffer resource, it results in a hang.

When a reader reads a buffer, it adds that buffer resource's wait
semaphore to its own list of wait semaphores, creates a new signal
semaphore, and sets the buffer's wait semaphore to its signal semaphore.
Now, if that same buffer is read a second time, this process is
repeated, and the command buffer takes the resource's wait semaphore
as its own, but since this time around the wait semaphore is actually
the signal semaphore from earlier, the command buffer now has to wait
on itself to fire before doing anywork, and so it stalls indefinitely.

By fixing this, an issue in the Snapshotter where snapshots would hang
if a snapshot is taken of a scene with multiple meshes that share the
same index buffer is also fixed (as the index buffer would be read
multiple times, resulting in the aforementioned issue).

Testing - Adds a new unit test to batch_gpu_uploader_unittest.cc.

Fixed: 39346

Change-Id: I3dcd33c429d521fbb353dfc2b7090ac5f5e315f7