[netstack3] New tool, ping3
The old ping binary is made in C and uses BSD sockets for sending ICMP
echo requests. This will not work for Netstack3, as BSD sockets are not
yet implemented and Sockets 2.0 is around the corner. The new ping3
binary uses the new ICMP echo socket API to send ICMP echo requests and
receive ICMP echo replies.

This CL will include a minimal implementation for ping3 specialized for
IxANVL testing.

```
$ chrealm /hub/r/netstack3-env/1767082
$ net fwd add-hop 192.168.4.1 0.0.0.0 0
Added forwarding entry for 0.0.0.0/0 to 192.168.4.1
$ ping3 8.8.8.8
PING 8.8.8.8 (8.8.8.8) 56(84) bytes of data.
64 bytes from 8.8.8.8: icmp_seq=0 ttl=64 time=3.656 ms
64 bytes from 8.8.8.8: icmp_seq=1 ttl=64 time=4.103 ms
64 bytes from 8.8.8.8: icmp_seq=2 ttl=64 time=3.636 ms
```

Bug: 21543 #done
Change-Id: I9943d3776410e0fb42032dc4c59f3335fcd49bfc