[pseudo-fs-mt] Allow CREATE{,_IF_ABSENT}
A subsequent change will introducde processing for the creation flags.
It is totaly file for a connection to carry those flags, so we should
not be returning errors for them.  OPEN_FLAG_CREATE and
OPEN_FLAG_CREATE_IF_ABSENT were not previously supported, so at that
point it made sense to limit their usage.

Change-Id: Ib5013dd0afb691be27a5816f2a090f87c4ac167f