[pseudo-fs-mt] MutableConnection: Entry creation
Tie up a number of changes from previous commits to allow directories
and and files to be created when a `OPEN_FLAG_CREATE` is specified in an
`Open` call.

It is still up to the user of the library to provide an
`EntryConstructor` that defines what entries are constructed and when.
This change just makes sure that the constructor will be used, when the
connection to the client is a
`directory::mutable::connection::MutableConnection`.

Change-Id: I8ca1a3159d1a3a9f6d45b87abfe7c28a7c4d5c1d