[scripts] FAR parser for Fuchsia Component Graph.
This tool provides the ability to parse FAR files into JSON. It can be
used both as a library and a standalone tool. This is part of a wider
changelist to introduce a component graph tool to Fuchsia. I was asked
to break this up into an individual CL to that it could be reviewed.

Change-Id: I334aab19741c3a8050ed1e2a3019f8cad631b88c
Bug: 38481