[netstack3] Make Icmpv6Packet a two-level enum
Migrate all of the NDP- and MLD-specific messages that were
previously variants in `Icmpv6Packet` to their own NDP- and MLD-
specific enums respsectively. Add `Ndp` and `Mld` variants to
`Icmpv6Packet`. This allows NDP and MLD code to operate on these
types directly, and thus eliminates a source of runtime panics
(namely, panicking if a received ICMPv6 packet is not actually
an NDP or MLD packet).

Test: No behavior change

Change-Id: I31409d4da088fecc93a4307f228e6d2ee4cb40e2