[usb][irq] Gracefully handle a lack of profile.
If we are unable to obtain a scheduler profile to apply to the high
priority IRQ completer, print a warning and continue to run.  Do not
fail to print _any_ warning and then fail to bind.  This is operating
on the principle that "it is better to just do the best you can with
what you have than give up and do nothing at all".

Change-Id: If72bc14dab884533f7b66b02575c07015fc3d806