[fuchsia.netstack] generic FIDL tests for N2/3
Made the tests in src/connectivity/network/tests/src/lib.rs generic over
Netstack2 and 3 so they can be run for both. Converted some tests to run
against N3, the rest of the tests can't be converted yet because
N3 currently needs interfaces to be up before addresses/forwarding
entries can be added/deleted and there is no good way to block until an
interface is up. This will be resolved through issue 38311.

Test: fx run-test netstack_integration_tests
Bug: 37717
Change-Id: I022f3d92de282aa99cdeb81873f612e3d38907fe