[debugger] Track register and constant source locations.
Tracks register source information from the direct register retrieval in
expressions ("print rax") as well as the DWARF expression evaluator.

The DWARF expression evaluator needed to be enhanced to track the
current register if there is one. It also tracks if the result is a
constant value. This allows us to generate nicer error messages if you
try to assign to a variable with a constant value.

This does not yet do anything with the resulting locations so there
should be no user-observable change other than error messages become
more clear.

BUG=39589

Change-Id: I006d60a3ed8505574f4604a78a5100c488a5d645