[libc] put initial handle extract in its own func
Currently, we have three locations for extracting handles:
- __libc_extensions_init
- __libc_startup_handles_init
- __libc_start_main

The first two are standalone functions but the last in embeded in the
middle of start_main.

fidlcat needs to know the semantic of the starting handles to be able to
display a useful information. For example, using the starting handles, we
are able to display:
zx_channel_write(handle:handle: ec6d7f4b(dir:/svc/fuchsia.sys.Launcher), options:uint32: 0)
  sent request fuchsia.sys/Launcher.CreateComponent = {
  ...

(Here we give the semantic of handle ec6d7fb4).

To be able to harvest and infer this, fidlcat puts a breakpoints on the
functions which extract handles and reads the memory adressed by the
arguments. This is only possible for standalone functions.

This patch puts the handle extraction which currently is directly in
__libc_start_main in its own function (processargs_extract_handles).

This way, fidlcat can put a breakpoint on it.

The overall behavior is unchanged.

Change-Id: Ia196173a160b0061f628aa11dd84a447bc40519a