[KMS] Add get_crypto_provider_name API for keys.
This allows user to check the crypto provider name backing up a key.
Since we now support all the operations for both software/optee
provider, we can turn on the test for all providers. In order to
verify the actual provider we are testing against, we add the
provider_name_test. I would add the name test for SoftwareProvider
in a later CL.

Tests: Unit tests and integration tests.
Change-Id: I0461074e0075952a55ee064ed1f29eaf59213a3c