[media][tests] Move 24kHz to out-of-band range
This minor test-only change moves 24 kHz from the set of "in-band"
frequencies, to the set of "out-of-band" frequencies. For this reason,
Because it is now considered out of the expected frequency range, we
no longer save previously-measured results for frequency response and
phase response, so these arrays have changed.

This CL prepares the way for the upcoming change to move the "brick
wall" low-pass filter slightly lower in the frequency spectrum--to a
more realistic audibility limit of perhaps 20kHz.

Test: test-only change

Change-Id: I90cd13add5560beefa150e6efde288aaba04567c