[fidl] compatibility_test support for xunion migration
Build two versions of the FIDL compatibility test servers for HLCPP and
Go and add them to the test matrix. Remove Rust and LLCPP from the test
matrix - they should be reintroduced as they gain read and write support
for unions as xunions.

This test demonstrates that servers written against the HLCPP and Go
FIDL bindings can interoperate if they can read unions as xunions,
whether or not they're writing unions as xunions.

Change-Id: I400a1212759d3bb27f2d787ea1b695b1a17b6521