[gidl][go] Generate tests for old and v1 wire formats
This is the first step towards encoding/decoding unions under both the
old and v1 formats in GIDL. When using the syntax for multiple wire
formats (added in fxr/321906), GIDL will now produce differently-named
test cases and pass the wire format information down to the backends.

This CL only implements the necessary backend changes for Go, so the
simple test case added in conformance.gidl is currently limited to Go.

This CL also renames the wire format constants from FidlWireFormat and
TestWireFormat to OldWireFormat (the current one, the default) and
V1WireFormat (the new one, where unions are encoded as xunions). This is
consistent with the naming used in the IR.

Test: fx build host-tools/gidl && fx exec $FUCHSIA_DIR/tools/fidl/gidl-conformance-suite/regen.sh && fx build third_party/go:go_fidl_tests && fx run-test go_fidl_tests -- -test.v

Bug: 39561
Change-Id: I4dbe25c2e782a0a3ae879b180f8dc8f5aa46215a