[amber] Add more logging when dealing with encrypted blobs
If amber is setup to expect encrypted blobs, but the repo blobs aren't
encrypted, it is a little hard to diagnose why package installation is
failing based off the logs. This adds some more logging messages to
show why amber is refusing to install a package.

Test: no functional changes. In repo.go, err is always nil at the
end of the function, so this also removes an unnecessary conditional.

Change-Id: Ia44e7057a14e9811171a69c860e81ed4c8f80d96