[camera] Fixing frame tracking in DmaMgr
The ArmIspDevice used to keep track of the first
time the dma would run, to have it signal that the previous
frame had been written.  Now the DmaMgr keeps track of this
variable, which may differ between the two dma's.

Testability:
This is tested by the modules_bin test, as well as the mali_009_test
For more information about how to test the camera on-device,
see //src/camera/TESTING.md

MULTIPLY: `[
  {
    "target": "//src/camera:camera_full_test",
    "os": "fuchsia",
    "total_runs": 100
  },
  {
    "target": "//src/camera:camera_full_on_device_test",
    "os": "fuchsia",
    "total_runs": 100
  }
]`

Change-Id: Ie2a12c2c6f92c6e469dd6bbb8b8d88f24ce94f82