[alln] Script for fixing compile errors by adding a config
This can be used to automate some of the work to harmonize compiler
flags by just finding all the targets that fail with the
introduction of a new flag and opting them out.

This script should find every target except where strict header deps are
missing. Once those are fixed though and we turn on `gn check` we should
be good. In the meantime the script usually automates >90% of the work.

Bug: 3203
Change-Id: I73e0b4e58d117a50fc027415db6dfe879cfe3803