[brcmfmac] Clean up firmware loading
This commit replaces the Linux-ported firmware loading code, which is
six layers of asynchronous callbacks that are not actually being used
asynchronously, with a much simpler straightforward version.  It also
replaces the existing state-machine-driven NVRAM parser with a more
robust and tested iterative version.  In all, the firmware load code is
reduced by more than half.

Change-Id: I4f9d61a23e9f4ea2873ddefaab7cfc0323950fae