[i18n] Adds ability to serve initial profile
This change adds a few flags that allow intl_property_manager to serve
a nonempty initial profile based on the flag settings.

If asked to serve an initial profile, the server will also log that
instruction to syslog.  A simple component that serves static profile
can come in handy in testing and debugging.

Bug: 38096
Change-Id: I2ff88a63a34a831ce0d548d35d12c0a87f83dedf