[debugger] JITD Attaching.
When the client issues an attach command, the agent now will look if
that process is waiting on the limbo. If so, it will retrieve it. If it
cannot do that, it will fallback to normal process attaching.

This permits the agent to retrieve the exception a thread is in.
Otherwise, the debugger will attach to that process, but it won't have
the exception handle.

TEST=Added unit test.
     Tested manually.
     fx run-test debug_agent_tests

Change-Id: If327374bdfa061b33f1dd48d3aa83dbd831d130e