[kernel] Rework job policy
This CL does not change behavior but it introduces
the concept of policy override

ZX_POL_OVERRIDE_ALLOW
ZX_POL_OVERRIDE_DENY

In the future this allows for a child job to
override a given policy if such policy allows it.

This was possible before with the construction
of as many jobs and overrides were needed but it
becomes unwidely.

The main user of this feature is the incremental
enablement of the bad handle policy in select
part of a job subtree.

As before, policy is frozen once a job has
children.

Other improvements
- It now uses fbl bitfield, easier to reason

No change in behavior, existing tests suffice,
the tests are in 3 places:
- k ut all  (JobPolicy object)
- test/sys/policy-test
- core-tests  (policy and slack tests)

Change-Id: I1d83a86001ed837eaf4bef4b8dc0036ff78b0c8f