[build] Remove use of symlinks for asan-variant host executables
This is a revert of d2e0727143742e3bc5a9cecb82aa1797ce122297 ("Use
symlinks for host_toolchain variant binaries"), but with a comment
added.

This is in preparation for converting performance tests in fuchsia.git
to be run via SL4F.  Running performance tests via SL4F involves
invoking catapult_converter as a host-side tool.  That involves
copying catapult_converter into a "runtime_deps" directory.  Before
this change, the catapult_converter inode that would get copied is a
relative symlink, but that becomes a dangling symlink when copied to
the runtime_deps directory.

The problem can be observed with an earlier version of this change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/329392

This config reproduces the problem:

  fx set terminal.x64 --args='select_variant=["host_asan"]'

That is the config used on the "terminal-x64-release" builder, which
enables "host_asan".

It produces the following dangling symlink:

$ readlink out/default/host_x64/gen/src/tests/end_to_end/perf/runtime_deps/catapult_converter
../host_x64-asan/catapult_converter

That relative path was only valid when the symlink was in its original
location of out/default/host_x64/catapult_converter.

Bug: 3166
Bug: 23091
Test: trybots (including integration) + manual steps above
Change-Id: Ic899cc128a5ca5444096f22cd59188beb872b913