[graphics][compute] Switch to static library targets
Required for rust bindings to link to spinel code. Note that
each static library was also given a unique name, which is
also required by rust bindings.

spinel_target.gni was updated to require a target name instead
of a config filename. Target name is used to determine the full
C99 config filename. This matches hotsort_target.gni better and
is needed to generate a unique name for this static library.

No change in behavior.

Test: fx build
Change-Id: I44a91d5096fedbaec4a9bbca1ae4517e686ef98e