[brcmfmac] Refactor BCDC protocol attachment
This commit moves the BCDC protocol implementation to its own build
target, included as a dependency only on those platform targets that
require them.  It also moves the protocol attachment point out of core
logic, such that each platform device will explicitly attach the
protocol that it requires.  This is in preparation for adding the MSGBUF
protocol, required for PCIE devices.

Change-Id: I4ec1755310a371cae1f3daca70c7a5477b26bb48