[bindc] Insert deprecated keys
The new bind language doesn't have any way to express keys that are
integers, but the bytecode doesn't have any way to express keys that are
strings. In the interim we can manually define the existing set of keys
inside the compiler so that bind libraries can extend then, and bind
programs can refer to them.

Bug: 34604

Change-Id: Ie3083d3c5b2705e1907cb4a187956758c3c8ac38