[fidl][c] Start of fidl_transform
Introducing a FIDL wire format transformer, capable of converting
encoded FIDL messages between formats.

This change introduces the framework for transformation, as well as most
of the implementation to convert from the v1 wire format (containing
only extensible-unions) to the old wire format (which has
static-unions).

This transformation will be used in the LLCPP bindings in order to
flexibly switch the on the wire representation, while maintaining in
memory object layout.

Test: transformer_tests.cc contains extensive tests
Change-Id: Iaeb1876d718c1fd4aff7bf4023e25729fcf3bc13