[cobalt] Test build setting to capture log data locally.
This CL makes it possible for testers of Cobalt to capture log data
locally on the device, so they can inspect it and run it through Cobalt's
pipeline manually. This is done by setting the build argument to LOCAL:

fx set [product].[board] --args='cobalt_environment="LOCAL"'

This build argument will never be set to LOCAL for any non-test users
or use cases.

Tested: There is no functional change in this CL for non-test use cases.
The refactor of the cobalt_app.cc constructor to enable the new testing
continues to be tested by the cobal_testapp_no_network integration test.

Bug: 3896

Change-Id: Ife55d9196abbd6d423254c34c05cab9282280ad1