[cmc_json] - Add rights definitions to the cm_scheme and rust bindings.
This change adds optional rights properties to the use, offer, expose
directory route declarations.

This change also defines a Rust structure Rights that is an alias for a
Vec<String> which maps to the set of expanded rights (no right aliases
in the .cm format).

Passing these directly through from compile.rs will be handled in
another CL.

Bug: 36541
Change-Id: If267873da2e3f7525e91353bf4bebd7a20301c02