[tools] - Component Graph.
The Fuchsia Component Graph maps components into a graph structure to
help audit and visualize how component services are connected. It
requires fx serve to be running which it uses to retrieve all required
metadata.

One caveat is that some services are built in and are not distributed as
packages. I have provided a builtins.json which requires manual
annotation of these services to form the true connected graph.

Change-Id: I808eb2716c857e977ed6374db92cdb936dd40007