[cm] Remove unused "instances" key.
We currently insert new RuntimeDirectory objects into the "instances"
map, using a random number as a key. We don't yet use this key (though
may have future plans), but risk the RNG generating the same key causing
the new instance to be dropped prematurely.

Remove the as-yet-unneeded key.

No change in behaviour.

Change-Id: I79029bb6fb0c5f19be1cdb6a8fc4bc698ff23161