[modular][basemgr_launcher] Reland Allow specific arguments
Relanding fxr/304428. There was a bug in modular config xdr parsing that
caused ledger not to initialize properly with this change.

This allows basemgr_launcher to accept a base_shell url as an argument for
benchmark testing via shell scripts.

This also changes basemgr to hold onto a copy of the Modular
configuration and host it for sessionmgr. Previously, basemgr was cloning
the directory it was reading configs from and passing that along to sessionmgr.
This doesn't work when basemgr_launcher starts basemgr and dies before
sessionmgr is initialized because the config directoy won't exist.

TEST=fx run-test basemgr_launcher_tests

Change-Id: I6b39adaf0b2cbe867e99966e66629d805773c7d6