[dhcpd] Implement ServerDispatcher.dispatch_{get|set|list}_option methods
This change implements the get set and list option methods of the
ServerDispatcher trait for Server. These methods correspond to the
similarly named methods of the fuchsia.net.dhcp.Server service, which
provides for retrieving and configuring DHCP option values from a DHCP
server. In order to implement the aforementioned methods, this change
also implements several conversion traits for the runtime equivalents of
the fidl values manipulated by the above methods.

Issue: 20941
Change-Id: I1f764f45eb5387e82bd9d0e68e9509f68fdf7f55