[dev][hid-input-report] Move LLCPP builders to class
Move the LLCPP builders from the stack into the
class memory. With the creation of touch,
these are actually a sizeable amount of memory
(~50kb) and they cause an error when they are
allocated on the stack.

Also decrease the max amount of reports that
can be sent at a time from 128 to 50, which
effectively halves the amount of memory
we need to keep in the driver.

TEST: Covered by hid-input-report-test
Change-Id: I556575e5b21159edef1d6de864019e03fb0b9f88