[scripts] Re-add d3.v3 to temporarily support legacy js in blobstat.
This change re-adds d3.v3.4.4.js (released in 2014) to the
//script/third_party directory to temporarily support the
fx blobstat utility as it transitions to d3 v5. It is versioned
as an older version as to not collide with d3/ which holds the
latest version of d3.

They cannot occupy the same directory since they have slightly
differing license terms. This is a temporary solution intended to allow
porting the fx blobstat tool to the latest version.

I have filed bug: 39875 to track the migration.

Bug: 39875
Change-Id: I9b6d9915ada431dc42acdb8654487b1d20972ea7