[debugger] Send back latest values after reg write.
The client caches register values for stack frames. When it writes a
register, it needs to update this cache.

This change sends back the new values from the agent as re-read from the
kernel. This ensures that the cache is up-to-date and makes sure there
are no mismatches: only data from the agent is used to update the cache.

BUG=39589

Change-Id: I8ea080fef8661fe60484da57d7790a8c4bfbd6d2