[kazoo][syscalls] Remove syscall reordering to match abigen
Previously, there was a hardcoded list of syscall names in kazoo which
served to reorder output so that it would textually match the output
produced by abigen/banjo. Now that we've moved off abigen this is no
longer necessary.

While the generated files have large diffs, no functional change is
intended.

This was previously only tested by the old "difftest" script, but
MakeSyscallOrderMatchOldDeclarationOrder() wasn't called in test code,
so no tests change -- i.e. this is deleting now-unnecessary untested
code.

Bug: 39918
Change-Id: I64e9f798580e2a7b80de36e820248cda53e69707