[kazoo][go] Add support for Go syscall emission to kazoo
This is a relatively rough conversion of mkfuchsia.go
https://fuchsia.googlesource.com/third_party/go/+/refs/heads/master/src/syscall/zx/mkfuchsia.go
that handles the same three outputs that it deals with.

Contrary to other output generators that were moved from banjo/abigen to
kazoo, this one is not identical to the .go version. Specifically:
- All syscalls are emitted rather than a hardcoded list
- All are output as Sys_ rather than a combination of sys_ and Sys_.
- Additionally, it looks like the generator hadn't been run recently, so
some names have been updated, and recently modified syscalls are changed.

No intended behavior change, however, this tool is run manually so the
outputs won't be used until we do so by running it inside
//third_party/go and rolling that in any case.

A subsequent change will deal with mkfuchsiavdso.go's outputs which are
not addressed here.

Unittests added.

Test: zircon/tools/kazoo/go_syscalls_output_test.cc added, kazoo_test.
Bug: 33668
Change-Id: Idc7435d35b119ebf8fc29a35d694fb4451deb79b