[kernel][mtrace] Make PmuHwProperties a standard-layout type
PmuHwProperties is an alias for either X86PmuProperties or
Arm64PmuProperties, depending on the architecture. This type is return
by zx_mtrace_control.

Prior to this change the type was defined using
inheritance. PmuCommonProperties served as the base class for the
derived types X86PmuProperties and Arm64PmuProperties. Because of the
way these types were defined, X86PmuProperties did not conform to
C++'s StandardLayoutType so it was at risk of changing layout when we
change compilers.

This change removes the inheritance and redefines X86PmuProperties and
Arm64PmuProperties to contain an PmuCommonProperties field named
"common".

Callers of zx_mtrace_control have been updated to use the new struct
definitions.

Bug: 30802
Fixed: 39742
Change-Id: I0cff634a11e687aff091b726d4ba9791fa13f26d