[fidl][llcpp] Undo the conditional encode/decode optimization
This will unblock integrating the transformer into LLCPP. The
optimization as it stands today is a bit too dangerous, and we should
add it back only if we manage to precisely determine which types really
need coding in the future.

As an example, a struct with a strict enum still needs coding tables,
because the enum need to be validated against a defined range of values.

Bug: 40027

Change-Id: Ia500bd3df6df71c6b7ab2ed903514b57c06c16c1