[debugger] Make a shared AddressRange class.
Previously we had a simple AddressRange struct in debug_ipc with some
random helper classes, and then a more complex AddressRange in
zxdb/common.

This moves the more complete implementation to debug/shared where it can
be used by the protocol, the debug_agent, and zxdb, and updates users
accordingly.

Change-Id: I57b60e5f1dfd24498e6746c8344976b16dfbcecf