[FDF] Enable proper driver rebind
At the moment, the device_rebind api, is less than ideal as it returns
after issuing the unbind call. The api does not ensure that the unbind
is complete and will not make sure the device is bound. It also only
supports autobind, which many drivers that want to use this api like fvm
driver do not support.

Rebind is a 2 fold operation of unbind followed by a rebind.
Without both these operations complete, there is no guarantee that
the device is bound and the clients will not be able to use the device
until a succesful bind.

This changeset solves these issues

1. It adds a fidl rebind op which is synchronous.
2. Also supports manual bind for rebind.

Test: runtests -t devhost-test-test

Bug:35617

Change-Id: I9a251c38afba0bd5ab613746548e4063409dcb70