[debugger] Support writing registers.
Hooks up the register mode of an ExprValueSource to actually make RPC
calls to write the registers in the agent.

Move some bit handling from bitfield.cc to the ExprValueSource so it can
be shared with the new register setting code.

Fix a bug where we forgot to serialize the registers in the protocol
encoders/decoders. Added a test.

BUG=39589

Change-Id: Ib744b58507c4081cdceaf4afc1ba0f19360172fa