[camera][hwaccel] Support to set (change) output res.
Adds support to set (change) output resolution. On task Init, we pass
a table ot potential output resolutions (ImageFormat2's) that we can
switch between. This change adds an interface to switch the output res.
There is no callback when the output res change completes. Instead, we
return the current output resolution (in use) with every frame ready
callback. For GE2D, we need to re-allocate the canvas ids (since the
canvas ids depend on the height/width). For GDC, we need to associate a
config (vmo) with a output res (and pass this tuple in to task init).
This wil be addressed later (it is lopri).

Test: Adds unit tests for coverage of the functionality.
Change-Id: I87e1761b0686153a1eb39bd378e782f1ddea57f0