[debugger] Add support for rust "as" casting
The 'as' cast operation appears to be very similar to a C style cast.
It's stricter but most of the strictness seems not to apply in a
debugger context.

At just about any precedence this tends to look odd in expressions, and
with our type parsing handling C++ types, some expressions that should
be valid (but very weird looking) aren't because the type parser gobbles
operators.

For the moment, though, this should be fine.

Fixes: 6001
Change-Id: I7e4a927ab2f23a51c099491caaa904e059b8fb90