[modular] Expose intl::PropertyProvider to agents and runners
Agents and Runners must have access to intl::PropertyProvider.
Due to /data path requirements in existing devices under test,
agents must be launched in the /sys/ realm (using the
"sessionmgr_context" environment), but also launched with optional
command line arguments via the ArgvInjectingLauncher.

This is different from mods that are launched from the
"session_environment_", with "ArgvInjectingLauncher", so we
need two different instances of ArgvInjectingLauncher: one for Agents
and one for Mods.

Other services (like Ledger) launched by SessionmgrImpl will use the
launcher from the sessionmgr_context_ *without* the
ArgvInjectingLauncher wrapper.

Now that services from the /sys realm are no longer automatically
inherited (as of fxr/329390, the story environment is no longer created
as a child environment of the session), the
fuchsia::intl::PropertyProvider also needs to be explicitly exposed to
launched components.

Fixed: 16285

Test: fx run-test modular_integration_tests
Change-Id: Iacc1bb746ca76c423288c7b26ddc3ced9ce6c556