[FVM]Modify FVM tests to use rebind api
Currently FVM tests use the following sequence:

1. block rebind fidl api which will call device_rebind
libdriver api.
2. device_rebind schedules unbind on all the children but
does not guarantee whether the device is actually unbound and bound again.
3. The FVM tests, use Bind device controller api to bind the driver again.

In the current scenario, depending on the timing of these operations, #3
will fail with ZX_ERR_ALREADY_BOUND if #2 is still in the process of unbinding.
In the case where #2 actually completes unbind #3 will succeed.

This change does the following.
1. It uses the new rebind device controller api, which takes care of
unbind and bind synchronously.
2. Because we are using the rebind api, we do not need the additional
bind
3. The Rebind api has a strict requirement of clients closing the
children before callign the rebind on the parent. FVM tests are modified
to do this.

Test: while /system/test/fs/blobfs-integration-test; do true; done;
while /boot/test/fs/fs-management-test; do true; done;

Bug:35617,Bug:39460

Change-Id: Ib0ce3e9ef77a871f522f449630c23609f457d4ba