[debugger] Sort registers before printing.
Previously the registers were in the same order as whatever order the
backend inserted them in. This isn't always the order expected by the
user.

This change sorts the register by ID so they're in a consistent order
and there is a central place where the register ordering is maintained.

Change-Id: If4faab304da9f4e9c4645ae7dd6db354c6098ca8