[fidl][rust] Support encoding/decoding unions as xunions
This CL teaches the Rust bindings to conditionally encode/decode unions
using the xunion wire format, needed for the union-to-xunion migration.
This includes compiler-generated result/error unions for error handling.
On the encode path, it is disabled by default. On the decode path, the
mode is determined by a flag in the transaction header.

Previously fidl::encoding made it easy to decode FIDL objects directly,
without a transaction message wrapping them. With these changes, a
Context object is now required (specifying how unions are encoded).
There were a few places in Overnet and Ledger decoding without a header;
this CL makes them explicityly provide a context with the flag set to
false. This will not impact the migration because neither Overnet nor
Ledger use static unions in their interfaces.

This CL also adds support for multiple wire formats in the Rust backend
of GIDL, to enable testing these changes in GIDL.

Test: fx build host-tools/gidl && fx exec $FUCHSIA_DIR/tools/fidl/gidl-conformance-suite/regen.sh && fx run-test rust_fidl_tests

Bug: 38647 (Rust encode)
Bug: 38451 (Rust decode)
Bug: 39561 (GIDL old/v1)
Change-Id: I2c31a453deb642b3a20501eec366478948a1b97f