[session] Update directory structure
This updates the directory structure to match the Fuchsia Source Code
Layout guidelines.

The structure is as follows:

session
  bin
    component_manager
    session_manager
  examples
    elements
    graphical_session
  fidl
  lib
    element_management
    realm_management
  tests
    session_manager
  tools
    session_control

Each child directory of //src/session contains a BUILD.gn file with two
required groups: "all" and "tests." The exception to this rule is the
tests directory which only contains "all."

This also simplifies a lot of the development instructions (e.g.,
simpler `fx set` invocations).

Change-Id: I370e2a8f09db7c9ec03258d7826fcb9619df1597