[kernel] [uart] Finish reading data from the UART
This patch fixes the amlogic_s905 UART driver.
Previously; if we ran out of space in the cbuf
(got a whole bunch of characters at once) -- we would
break out of the read loop, and never return. This patch
resumes the read operation if it was previously interrupted,
in the context of the caller of s905_uart_getc.

Test: Disconnect the USB serial adapter from the host
and re-connect several times. Validate that serial
input and output still function after doing this.

Change-Id: I019dc3862c38efaac47ca61569f0502cd2715364