[benchmarks] Extend SL4F wrapper for benchmarks to produce output for dashboard
Extend the recently-added SL4F wrapper for zircon_benchmarks so that
it produces output in the format required for uploading to the
Catapult performance dashboard.

This involves adding an invocation of catapult_converter, which gets
run on the host side.  (The older wrapper around zircon_benchmarks is
in peridot/tests/benchmarks.  In contrast, that runs
catapult_converter on the Fuchsia device side.)

We disable uploading to Catapult until we can test that better via
trybots and until the other invocation of zircon_benchmarks is removed
from the peridot-perf bot (otherwise we'll get duplicate results
uploaded).

For local testing:

  # Use host_asan to match the fuchsia.git bots:
  fx set terminal.x64 --args='select_variant=["host_asan"]'
  # fx build + re-pave
  fx run-e2e-tests perf_test

Bug: 23091
Test: local testing steps above + the terminal-x64-release CQ trybot
Change-Id: I487fadadb006fa131203257d442c315a1195e045