[graphics][compute]: Inject tracing.provider.Registry into test
This CL ensures that test components generated for the graphics
compute project on Fuchsia inject the tracing.provider.Registry
service into the hermetic test environment.

This is to remove potential flakes that appear when the
graphics_compute_vk_unittests are run on NuC devices. See bug
for details.

Note that the test doesn't use tracing at all. It is likely that
the service is probably a dependency of the Intel Vulkan ICD library
instead which is loaded indirectly by the vulkan.loader.Loader
service on these devices.

Bug: 39889
Testing: fx run-test graphics_compute_vk_unittests
Change-Id: Ic69b55119790805fb1c56337b82e569b9c79ee95