[kazoo][syscalls] Rename protocols to snake_case
Previously

  protocol Channel {
    channel_create(...) -> (...);
  };

For consistency with method names, protocol is now lower case too.

  protocol channel {
    channel_create(...) -> (...);
  };

The protocols were only used to confirm that they matched the prefix on
the method, so no tests change.

This required a couple minor renames of "enum clock" and "using futex"
in order to not collide with the new protocol name, but nothing that's
currently externally visible.

No behavior change intended.

Bug: 33668
Bug: 39918
Change-Id: Ic537a5da537ed4e7939aeb79108ec785b9e7ceb4