[media.sessions] Fix potential flake
It is possible that a channel can close in one process, and at a later
point in wall clock time another process can successfully send on the
channel before realizing it is closed.

This test depended on send failure to check whether a channel was closed
which occasionally hit this race. This has been seen locally but not on
CQ yet.

MULTIPLY: `[{
    "target": "//src/media/sessions:mediasession_tests",
    "total_runs": 100
}]`

TEST: fx run-test mediasession_tests
Change-Id: Ie562b3da838fd0a2208bd4f93ed1041b0339c579