[runtests] Generate scripts at build time
The current test approach for runtests-utils involves dynamically
writing scripts on minfs, and then executing them.  That's not something
we want to support under our verified execution plans.

Instead, we can generate these scripts at build time, and then make use
of them out of the test root and test data directories.

This change allows us to remove the hack that we added to make this not
block landing the default-deny policy for replace-as-executable, so we
do that too, to prove that it's not still load-bearing.  Cleanup!

Bug: 37529
Change-Id: Id4358d72449f1f35b3bd1ce5b07d7bb33898ab92